# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VENDORA API",
        "description": "API documentation for VENDORA.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Application base URL"
        }
    ],
    "paths": {
...
        "/api/payments": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "List payments",
                "operationId": "afcbc19a07bf33c2fcb73b5c438b476c",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "method",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "payment_number": {
                                                        "type": "string",
                                                        "example": "PAY-001"
                                                    },
                                                    "order_number": {
                                                        "type": "string",
                                                        "example": "ORD-001"
                                                    },
                                                    "customer": {
                                                        "type": "string",
                                                        "example": "John Dela Cruz"
                                                    },
                                                    "paid_at": {
                                                        "type": "string",
                                                        "example": "2026-01-10 14:30"
                                                    },
                                                    "amount": {
                                                        "type": "integer",
                                                        "example": 2450
                                                    },
                                                    "currency": {
                                                        "type": "string",
                                                        "example": "PHP"
                                                    },
                                                    "method": {
                                                        "type": "string",
                                                        "example": "cash"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "completed"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 248
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Create a payment",
                "operationId": "985b848b4cb6ca6115cf4ea71e314c0c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id",
                                    "paid_at",
                                    "amount",
                                    "method",
                                    "status"
                                ],
                                "properties": {
                                    "order_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "paid_at": {
                                        "type": "string",
                                        "example": "2026-01-10 14:30"
                                    },
                                    "amount": {
                                        "type": "integer",
                                        "example": 2450
                                    },
                                    "method": {
                                        "type": "string",
                                        "example": "cash"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "completed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Payment created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "payment_number": {
                                            "type": "string",
                                            "example": "PAY-001"
                                        },
                                        "order_number": {
                                            "type": "string",
                                            "example": "ORD-001"
                                        },
                                        "customer": {
                                            "type": "string",
                                            "example": "John Dela Cruz"
                                        },
                                        "paid_at": {
                                            "type": "string",
                                            "example": "2026-01-10 14:30"
                                        },
                                        "amount": {
                                            "type": "integer",
                                            "example": 2450
                                        },
                                        "currency": {
                                            "type": "string",
                                            "example": "PHP"
                                        },
                                        "method": {
                                            "type": "string",
                                            "example": "cash"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "completed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payments/summary": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "Payment summary cards",
                "operationId": "d7aa2d3f3b85c9cd0a72c145aa20250e",
                "responses": {
                    "200": {
                        "description": "Payment summary",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_revenue": {
                                            "type": "integer",
                                            "example": 125450
                                        },
                                        "cash_payments": {
                                            "type": "integer",
                                            "example": 45200
                                        },
                                        "card_payments": {
                                            "type": "integer",
                                            "example": 58750
                                        },
                                        "online_payments": {
                                            "type": "integer",
                                            "example": 21500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}